adding verifications on controllers at methods update, show, delete

+++ backend/src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactsRepository = require('../repositories/ContactRepository');
+const isValidUUID = require('../../utils/isValidUUID');
 
 class ContactController {
   async index(request, response) {
@@ -12,6 +13,11 @@ class ContactController {
   async show(request, response) {
     // get one record
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Ivalid contact ID' });
+    }
+
     const contact = await ContactsRepository.findById(id);
 
     if (!contact) {
@@ -27,18 +33,27 @@ class ContactController {
       name, email, phone, category_id,
     } = request.body;
 
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Ivalid category' });
+    }
+
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactExist = await ContactsRepository.findByEmail(email);
+    if (email) {
+      const contactExist = await ContactsRepository.findByEmail(email);
 
-    if (contactExist) {
-      return response.status(400).json({ error: 'This e-mail already exists' });
+      if (contactExist) {
+        return response.status(400).json({ error: 'This e-mail already exists' });
+      }
     }
 
     const contact = await ContactsRepository.create({
-      name, email, phone, category_id,
+      name,
+      email: email || null,
+      phone,
+      category_id: category_id || null,
     });
     response.status(201).json(contact);
   }
@@ -50,23 +65,33 @@ class ContactController {
       name, email, phone, category_id,
     } = request.body;
 
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Ivalid contact ID' });
+    }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Ivalid category' });
+    }
+
     const contactExist = await ContactsRepository.findById(id);
 
     if (!contactExist) {
       return response.status(404).json({ error: 'Contact not found' });
     }
 
-    const contactByEmail = await ContactsRepository.findByEmail(email);
+    if (email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
 
-    if (contactByEmail && contactByEmail.id !== id) {
-      return response.status(400).json({ error: 'This e-mail already exists' });
+      if (contactByEmail && contactByEmail.id !== id) {
+        return response.status(400).json({ error: 'This e-mail already exists' });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
 
     response.json(contact);
@@ -75,6 +100,10 @@ class ContactController {
   async delete(request, response) {
     // delete one record
     const { id } = request.params;
+
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Ivalid contact ID' });
+    }
 
     await ContactsRepository.delete(id);
